+++ src/main/java/io/github/ferrazsergio/libraryapi/infrastructure/repository/LoanRepository.java
@@ -11,17 +11,17 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface LoanRepository extends JpaRepository<Loan, Long> {
+public interface LoanRepository extends JpaRepository<Loan, Integer> {
 
     @Query("SELECT l FROM Loan l WHERE l.user.id = :userId")
-    Page<Loan> findByUserId(@Param("userId") Long userId, Pageable pageable);
+    Page<Loan> findByUserId(@Param("userId") Integer userId, Pageable pageable);
 
     @Query("SELECT l FROM Loan l WHERE l.book.id = :bookId")
-    Page<Loan> findByBookId(@Param("bookId") Long bookId, Pageable pageable);
+    Page<Loan> findByBookId(@Param("bookId") Integer bookId, Pageable pageable);
 
     @Query("SELECT l FROM Loan l WHERE l.status = 'ACTIVE' AND l.expectedReturnDate < CURRENT_DATE")
     List<Loan> findAllOverdueLoans();
 
     @Query("SELECT COUNT(l) FROM Loan l WHERE l.status = 'ACTIVE' AND l.book.id = :bookId")
-    long countActiveLoansForBook(@Param("bookId") Long bookId);
+    long countActiveLoansForBook(@Param("bookId") Integer bookId);
 }
